requirements.txt
# Core dependencies
rich>=13.0.0
requests>=2.31.0
pyyaml>=6.0.0
jsonschema>=4.21.0
uvicorn>=0.41.0
fastapi>=0.135.1
sqlalchemy>=2.0.48
pydantic-settings>=2.13.1
psycopg2-binary>=2.9.11
passlib>=1.7.4

# Authentication & Security
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
bcrypt>=6.0.0
pydantic[email]>=2.0.0

# AWS
boto3>=1.34.0

# Azure
azure-identity>=1.15.0
azure-mgmt-storage>=21.0.0
azure-mgmt-authorization>=4.0.0
azure-storage-blob>=12.19.0

# GCP
google-cloud-storage>=2.14.0
google-cloud-resource-manager>=1.12.0
google-auth>=2.27.0
google-api-python-client>=2.0.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0